Create apple_tv_tvos13.markdown (#18)
I think this is probably worthy of an alert? tvOS 13 software will be released to the public within the next several days and it completely breaks the Apple TV media player integration for HA. Doesn't seem like there's an imminent fix on the horizon so it would be nice to warn people of this and be able to point people to this alert if their setup breaks.